Capture OS Type as part of the Photon-model enumeration
Overall this is needed ASAP for the UI to display appropriate Compute icons.
- Capture normalized Os Types as a ComputeState custom property.
- Linux vs Window only

Rather than using a custom property I would actually prefer a first class property since the goal is to capture this info in a normalized manner across clouds
(@George/Guru this is your call)

Also this change does not include automated tests yet. I will work on that next.

Change-Id: I41e247cf1b18fef6c048fdcc5a2f3fad77f4d737