Use custom body CollectionRemovalRequest to remove collections
- Use custom body:CollectionRemovalRequest to remove the elements
of Collection attribute which are no longer needed.
It supports multiple collections to remove in one request.
CollectionRemovalRequest will be used in the future for
the issue described below:

We want to remove the NetworkLinks which are no longer
needed. Current handlePatch menthod in ComputerService
only do the ADD. For detail, see:
https://jira-hzn.eng.vmware.com/browse/VSYM-718

- Add handle function in ResourceUtils to do the remove when the body
of PATCH is CollectionRemovalRequest type.

Change-Id: Ib5425e977994bc507c2de1aaa8fd09e861ead12b