Prune NICs if they are detached from a VM
During the updates to the compute state, if it is discovered
that one of the NICs is detached, the VMs will not have a public
IP any longer and the associated NIC will no longer be needed.

Further the corresponding NIC should be deleted from
the system as it is not being referenced from anyplace.

For details, please see:
https://jira-hzn.eng.vmware.com/browse/VSYM-718

Summary:
- Added methods to use custom body
ServiceStateCollectionUpdateRequest to remove networkLink
and its document.
- Added methods in TestAWSEnumerationTask to stop AWS instances.
- Added methods in TestAWSEnumerationTask to check whether the
stopped instance's NIC and its document has been removed.

Change-Id: I224fc3c4065fd545ffed70841835d7992fe62723