Deploy ovfs imported as ComputeDescriptions
When the computeDescription is based on an OVF the
instance adapter does an OVF import. While the OVF
property descriptions are in the ComputeDescription the
values need to be provided in the ComputeState.

The guestInfo OVF transport is enabled by default. This enables
the test to deploy a coreos and do cloud-config.

Change-Id: I6d09c0f8174951e01ee2b140e857ecb3d26fe224