Add historical rollup of metrics
Aggregate metrics are backed by ResourceAggregateMetricService
instances. Aggregate metrics are identified using a key that is
a combination of the resourceId and metricKey. This means
that the same service document is written to on every rollup
resulting in the documents being versioned over time. Queries
for aggregate metrics need to issue a query to include all versions
and filter by date

Aggregation operations can run multiple times within a time window.
This will result in multiple versions of the document for
the time window. The one with the highest version represent the
aggregate value after the time interval has closed. All
other versions represent point in time aggregates for the collection interval

All aggregate metrics have a timestamp that represents the end of the interval.
For example if the aggregation is for hourly data and the interval is 10-11, the
aggregate value will have a timestamp of 11

TODO:

- Aggregations across resources
- More testing
- Optimizations to compute aggregates off data in memory

Change-Id: Id3e159c9a03a4bf0d386e9950624e66328919baa