VSYM-1850: Stats Task and AWS Adapter changes to pass a window for stats collection.
Design :
https://confluence.eng.vmware.com/display/SYM/Stats+Collection+and+Persistence+Enhancements+in+Photon+Model
- A new in memory metric is introduced to save the value of
  the last collection timestamp. This is saved with key
"statsLink-lastCollectionTimeMicros".

- When the SingleResourceStatsCollectionTask needs to kick
  off a stats collection request to the underlying adapter,
it checks the value of the last collection time in the stats
endpoint of the compute.

- If the value is not found, then it looks at the value of
  the last persisted metric and uses that to set the
collection window for the stats request.

- A new field is added to the Compute Stats Request to
  record the timestamp of the last run that was successful.

- The stats adapter will use this timestamp to determine
  the window for which the stats collection should be run.
If this value is not specified OR is greater than the maximum
configured window size then the stats adapter
collects stats for the default configured window.

- Enhanced the existing test, to send a stats request to the
  AWS stats adapter for some intervals in the past. Since,
the machine is newly provisioned, it takes some time for the
additional windows to become available. So we have a relaxed
check to assert that atleast >1 data point was received from
AWS.

- Had to modify the StatsCollectionTask to drop the interval
for scheduled task and reduce the number of computes. With the
original configuration, the operations were not completing
in time before the test exit-ed and was causing test failures
downstream.

VSYM-1845 Adding the ability to save the lastCollection time as
an in memory metric.

Change-Id: I9382aa5c8f513304d7b458a71e1f40dd6b366b3b