Introducing TagService
As discussed, key-value tags are not meant to replace resource groups.
In the future, resource groups may be defined through a tag-based query.

I had hard time trying to guarantee a unique TagState document per
key-value pair but eventually gave up. Having the tenantLinks in
the TagState makes it impossible since the same key-value can be
available to multiple tenants, hence we can end up with
multiple documents for the same key-value.

This lack of uniqueness will make tag-based queries more complex
but we can try to address this as we progress with using the
new service.

Change-Id: I1bf333464d57ab61abb0f5d94d0a316073d8ff05