Support for query-driven (elastic) resource pools
Resource pools seem to have had the idea of "elasticity" based
on a dynamic query instead of hardcoded links from resources to
the pool they participate in. The query field was there but never
used.

This change proposes a full support for query-driven resource pools.
The query is set from outside for elastic RP, and is auto-generated
(as before) for non-elastic RP.

In admiral we are going to generate queries based on tag criteria
entered by the user (where the actual tag criteria is stored in a
separate service linked to the RP service) so that newly added
resources with the appropriate tags get automatically added to the pool.

But the query-driven resource pool is not restricted to tags -
it works with any query provided to it.

Change-Id: I4b56a994d352f62188084efba33aa20691d5afde