PATCH fixed to return the full patched state in the response
Used to return the original patch state from the request, which was
not very useful.

If state is not changed, 304 is returned with the original patch body.
When accessed through a HTTP/REST client, the body is not available.
When accessed in unit tests, the body is there (xenon magic).
Tried with op.setBody(null) but it had no effect - in unit tests the
patch body was still there.

Change-Id: I37a6a075ba693c3486b31f23eaef8a6507a9eccd