Relaxing query validation when creating non-elastic resource pools
Non-elastic resource pools have their query auto-generated, and there
was a validation to make sure the client does not provide a query in
this case.
However this didn't work well when migrating existing resource pools
to another node (since the query is already generated and set in the
POST request).
Relaxed the validation to just ignore the passed query and
auto-generate it again.

Change-Id: Ia6cb5e86d7dbd2b728b67933cfe04bfe916d8518