Remove ComputeState.networkId
networkId is enatai legacy and makes no sense as a VM
can be part of multiple networks. The relationship VirtualMachine->Network
is expressed as:

ComputeState
  -networkInterfaceLinks->
      NetworkInterfaceState
        -networkLink->
          NetworkState

Only plain networks are supported for now (no DVS).

When cloning from template the networkInterfaces are also cloned and the
ones in ComputeState are ignored to preserve the "clone" expectations.

During enumeration entities are collected in a top-sort order
to make sure dependencies are available later. As a result
enumeration task for vSphere finish when all state updates are done.
Tests are simplified but enumeration is more sensitive to bugs.

Change-Id: I37e5fbad4007a2f522a0e977ad8bbcbb4aabb4e5