Optimize SingleResourceStatsAggregationTaskService
We now invoke one query to fetch all relevent raw metrics and
filter the resultset.

This results in one query being invoked per resource instead of
of around 20 for AWS resources. When invoked via the
StatsAggregationTaskService we will see 100 queries per batch
instead of 2000

We can further optimize this a few ways:

1. Invoke one query per batch instead of per resource. This
needs a change in how the aggregation tasks are constructed

2. Do not perform daily rollups based on raw values each time.
We only need the raw values for the open hourly interval and can
use the hourly aggregate values for the closed intervals. This
requires how we compute aggregates at various granularities and
will be done in a later changeset

Change-Id: Ie6d380a27f997532290a7e3def9dbf1f928d46d9