Enable dynamic placement
Compute's that can host a VM are now collected with their root
ResourcePool.
To control where a VM ends up the custom property __placement has to be
set to a link to a another compute that has a root resourcePool.

Hosts are not enumerated anymore. Instead they end up in the index as
non-clustered ComputeResources.

All IT tests depend on enumeration task to finish so they can pick a
resourcePool at random.

Introduce convention in naming customProperties constants.

Change-Id: I93cdd58937ad23eebe31f5557584702bb792eb81