VSYM-3136 : Fixing issues with endpoint deletion.
- Adding additional indexing options to the
customProperties in ScheduledTaskService.

- The deletion query for the associated documents in
  the EndpointRemoval task service was not able to
match against the ScheduledTaskService as the expand
option was not set this service. As a result the endpoint
deletion would leave behind the scheduled task for enumeration.

- Further removed the part that deletes the
  resourcepool associated with an endpoint.  The
resource pool is a project level construct and should
not be deleted during the deletion of an endpoint.

- Tested the change and validated that now the
  schdeuled task for enumeration is getting deleted
when the endpoint is deleted. Also checked that
stats collection at the resource pool level etc
goes through without the serviceNotFoundException
seen earlier.

Change-Id: Ic06af256d7e8d00f0f591e17f5d7e6a9016980f9