Fix vSphere host enumeration
We discover:
1. DRS-enabled clusters (w/o the hosts inside)
2. Hosts from a non-DRS clusters (w/o the cluster itself)
3. Standalone hosts.

For each of these objects we keep the resource pool that will be
used to provision VMs (this is a required property in the vSphere API).
For (1), we keep the root RP of the cluster.
For (2), we keep the root RP of the cluster, and during provisioning we
always specify the actual host (this is mandatory for non-DRS).
For (3), we keep the RP of the ComputeResource hosting the HostSystem and
basically treat (1) and (3) in the same way (as ComputeResource's).

Change-Id: Id9bb3171614af01b2018aa99a4f51004f0b74004