Part1: Change the ResourceRemoval behaviour.
- SubTaskService will now wait for all the completionsRemaining, before patching
back parrent task. It will patch the parrent task with a well defined patch body
ServiceTaskCallbackResponse, which will contain the state of the task and if any of the
subtask are failed a Set of failed ResourceOperationResponse.
- ResourceRemovalTaskService is responsible to delete local documents.
- Adapters must delete only remote resources.

In the second part I'll adapt the Adapters to respond to Task requests with ResourceOperationResponse.

Change-Id: Idb3784cbfb0d18f216a2036c488741c6be39e1b8