VSYM-2495: Add storage utilization metric for Azure disks
* Introduced new service that collects storage stats per compute host.
The current implementation stores the storageUsedBytes as a total of
storage that is being consumed by the blobs at the compute host level.
* Intorduced new service for aggregation of compute host level stats.
With this the flow of coumpute host level stats collection changes
slightly. From the AzureStatsService, for each host, we trigger the
AzureComputeHostStorageStatsGatherer and the AzureComputeHostStatsGatherer.
When they both return we bundle all the metrics in one response and
send it back to the SingleResourceStatsCollectionTaskService.

Change-Id: I59ecf848537e3c844ba444b27e3960d23fbef979