Mark photo model task services non-persistent
The only service that is marked persistent is the scheduled
task service. This is used in a few places to reschedule
tasks across restart. The semantics of that will be changed
to seperate the state data from the actual task so that
the task itself can be transient with just the state persisted
That is however a breaking change and will be done in a later
change

This change is in preparation for moving tasks to a different
xenon node group that can be upgraded independently

Change-Id: I68507161689653410f08a72f9f2148e56cf06837