Return the TagState on idempotent POST
On duplicate tag post (transformed to a PUT), we used to return 304
and the current state in the response body. This works in unit tests
but does not work through an http client - the body is always emtpy
for 304.

This is a problem for tags since a typical use case is to always try
to create the tag and then use the returned self link.

With this change, we now return 200 and the current state in the body.

Change-Id: I89b10a7d7e82af750d5e3c25b4624172acb13e31