Initial create NICs for AWS
The logic adds Nic Context structure to the allocation context, and for each Nic Context the NetworkState, SubnetState, FirewallStates are obtained, corresponding AWS objects are discovered, and AWS InstanceNetworkInterfaceSpecification object is created for each NIC. Those objects describe the NICs that will be attached to the provisioned VM.

In case the NIC number is 1, the instances will get public IP address assigned to their primary (and only) NIC based on subnet configuration, so that the behaviour will be exactly the same as before this change. However when assigning more than one NIC, public IP address will not be assigned.

Change-Id: I81318a922b999dec731ec4d00b24e53f5ae208c6