[vsphere] Garbage-collect compute resource after enumeration
It is possible that resources are removed between enumerations.
Depending on the value ComputeEnumerateResourceRequest#preserveMissing
computes are either makred as RETIRED or deleted.

The set of resources to be gc-ed is formed by querying all resources that have
not been enumerated during current enumeration or have never been enumerated.
A custom property holds a link to the Task through which it was last found.

Change-Id: I626158dc682a3271aac02b61b16e2d560a852863