[vsphere] Unify OVF & cloud-config properties
Just as with cloud-config the OVF props are now read from the BootDisk.

When FileItems are detected in the BootConfig they are tranlsated to OVF
properties. It is possible that the same cloud-config property end up as two
keys as there are two cloud-config implementations currently (coreos and
ubuntu). As a result, by just setting cloud-config properties in the BootDisk
clients can ignore OVF details if they are concerned only with cloud-config.

Change-Id: I3dc0fafd7fff25394af6489dc55346ef5fd979a6