VSYM-3689: Remove resource when deleting Azure or AWS endpoints
*Introduce endpointLink to services
*Resources that end up getting deleted:
AuthCredentialsServiceState, DiskState, ComputeState,
ComputeDescription, ComputeStateWithDescription, NetworkState,
NetworkInterfaceState, NetworkInterfaceDescription, SecurityGroupState,
SubnetState, StorageDescription, ResourceGroupState
*Associate every resource created as part of an endpoint enumeration
with the endpoint itself. The association can differ between resource
types. The properties used for the association (based on what's available
on the resource kind) are either endpointLink or
customProperties.endpointLink.
*Query for all documents that belong to one of the resource types listed
and have the above properties set to the endpoint and remove all results

TODO: TagServiceStates and ResourceMetrics are not removed as part of
 this change

Change-Id: I9be740356355dcaffcfa5d4c12f676e9be257777