Tag Azure NetworkStates with link to correct ResourceGroupState (via groupLink property)
Along with this fixed a problem when running more than one enumeration.
The problem is that if a state object is patched but actually no modifications were made then the documentUpdateTimeMicros is not updated and the enum adapters delete those resources. I am changing to a mechanism where a list of azure resources ids is used in the query to select resources that are not available there.

Change-Id: I38d4bd9eee18dd46e209d71d8675ca0ae744a720