VSYM-3864: Adding the environment name to compute state.
- This information is needed for the resource list
  view.
https://confluence.eng.vmware.com/display/SYM/Common+Resources+List+UX?preview=/171897710/176147890/Screen%20Shot%202016-10-05%20at%2012.04.30%20PM.png#CommonResourcesListUX-Listofresourcetypesandcolumns

- The resource list view is backed by a materialized view that holds
  data about computes and the associated metrics. Since
the environment name is not a first class attribute in the
Compute State ; deriving this information for the
resource list view would mean creating a materialized
view across more documents. Since that is more
expensive, resorting to the addition of a first class
attribute "environmentName" to represent the cloud that this
compute belongs to.

Change-Id: I9748fdc59467e30687f51b3acaf2ed8ac5a8b025