Modifying result limit for Photon Model queries.
- Setting the default result limit 100
for the TOP_RESULTS queries.

- Set the AWS Network Query result
limit to 1k as it is expected to
have a small number of documents.

- The result limit was set
to an egregiously low number determined by the number
of entities that existed on the remote endpoint.

- This was likely causing performance issues at
the Xenon level due to a very small result set
and many iterations over the data to arrive
at the final result.

Change-Id: Iceaf6c1ff7dc99ef55a2afab9b7de42b3de2152a