Last rollup time lookup optimzation
The last rollup time a metric was aggregated is obtained via a
query if there is no in memory stat available. This optimization
did not work for cases when a metric is not applicable for
a resource. This change adds an in memory stat for the key with
a value of 0 to tell the aggregation logic that it does not have
to lookup the last rollup time for the metric.

There is no functional impact with this change as we will
populate the stat with the right value if the metric were to
become available

Change-Id: Ib71f3de05d9d2f70a57a28bac199e235797ba2f1