[Azure]Fix validate virtual gateways
Fix for https://jira-hzn.eng.vmware.com/browse/VSYM-4192
Attach gateway to the shared resource group to only provision it once
and not on every enum test run. Also moved it to a separate test.
That works around the slow provisioning and decommissioning of Azure gateways that
intermittenly break preflight.

Change-Id: I9926d85222387216db36d71903d59826f239d327