Adding a contraints field to ComputeDescription
This allows specifying constraints of this compute against other
resources. The immediate need for this is declaring placement
constraints such as "place this VM on a hardware that has the
pci_compliance tag and preferably (soft enforcement) does not
have the location:sofia tag".

Implemented in a generic way so that other services can also add
their specific constraints on the desired state.

Change-Id: I4944be03f25ce299156d01457da734c4dbee885f