Delete subnets even if n/w has not been deleted
The current logic assumes that subnets will only be
deleted if the subnet associated with it has been
deleted in the current cycle. If the subnet creation fails
for any reason in that cycle it will never be deleted again

Also, there is no need to check for auth credentials service
trying to ascertain identity of the network. It is scoped
to the enppoint and tenantLink

Change-Id: I956f8311cf734c86aae35786fc9e9b9ecbdd34ad