[Azure] Delete VM is fixed to wait until VM RG is successfully deleted from Azure
In this CL:

1) AzureInstanceService.deleteVM is fixed to wait until
the RG is successfully deleted from Azure. This seems to be
time-consuming task: ~5-7mins.

2) TestAzureProvisionTask tests are fixed to tearDown correctly
through the Adapter.Delete, but not directly to delete the VM RG

As a result of 1) & 2) testProvisionVMUsingSharedNetwork is IGNORED since
azure build timeouts due to the time consuming provision-decommission-VM cycles
(~10 mins) executed by testProvisionVMUsingSharedNetwork + testProvision.
The test will be added again once we optimize provision-decommission-VM cycle
(through using shared storage account, for example) or
tuning signal-completion-upon-deletion strategy.

3) AzureDecommissionCallback is introduced by analogy with
AzureProvisioningCallback

4) RetryStrategy is added to control the retry period of Azure
provisioning and decommission logics

5) AzureTestUtil.deleteVMs is fixed to wait for ResourceRemovalTaskState
to complete

Change-Id: I2b22afbed9437f2b6e06627f04a52cd9f5be7423