Turn off PATCH auto merge of primitive fields in ComputeDescription
PATCH used to always replace (usually with zeroes) the cpu/memory figures.
Fixed to only replace if the patch has a non-zero value.
Now zero is assummed to be a valid value indicating unlimited resources.

Another way to fix that was to use boxed types but that would have been
a more disruptive change.

Change-Id: I921c6b793cfa50d9db00eab989831454a0482f72