[AWS] Enumerate tags in Networks and Subnets, Update BaseEnumerationAdapterContext
BaseEnumerationAdapterContext is now in sync with the approach,
used to enumerate CS, Networks and Subnets (described below)

BaseEnumerationAdapterContext does not use NAME tags to set the ResourceState name
this is because there are resources without Name tags (like SecurityGroups)
In case some dependent class needs his resource to be named over tags,
he needs to implement this.

1. Create TagStates for the new remote tags
2. Add tagLinks to the NetworkState and SubnetState (do not merge,
since TagService works with IDEMPOTENT_POST option and
- creates only new TagStates (with new key-value sets)
- create a new TagState for each new value for the same key)
3. For networks and subnets: In case there is Name tag, set the name as ResourceState's name

This behaviour is consistent with the one for ComputeState tags
Subsequent CL should update the BaseEnumerationAdapterContext to only
post tag states following the same approach.

Change-Id: I081a7867a198b0726ed64fa622d967434e5ad999