[AWS] Region Id should not be required during endpoint creation
The region Id is no longer required in the endpoint and 2 of the main
tests are changed to work without passing region Id when obtaining
the endpoint client.

Additionally in the CL are included 2 leftover small changes from my past
story - tagging during enumeration.
Rename of one method, and extract common code to a method.

Change-Id: Ib2082b207595a92398e8537e7bd125ee06f4a430