Distinguish local vs. external tags
'External' tags are those imported from cloud providers.
If they get removed from a resource on the provider side, we
remove them in the local state too.

Once a tag is defined in local operation (e.g. defining a constraint
for placement, for example "all hosts tagged with pci_compliance")
the tag is marked as 'local' (even if it used to be 'external' before).
This is to ensure that remote changes will not affect the local view
that the user has validated and relies on.

If 'local' tags are unassigned on the provider side, they will not be
automatically unassigned from the local state.

Change-Id: I43a0ca23e5db5b73c45537d4914563d5d1cbdba2