[Azure] Fix Enumerate update compute state tags
The test used to set a number of tags and check for the created
resource states after 2 subsequent enumerations.
Now the tagging is moved to after the first enumeration
so that it covers the fix in the EnumerationAdapter.
TODO add assert for tag states number in the test
Current change adds logs to identify where the the instability comes from,
because the check for tag states
is done directly after the tagging and enumerating with
latest state.

Change-Id: I90e9aa2d8f05874136992ae50b8dcebf6183f2dd