Fix AWS preflight
The id of a security group has changed (probably the security group was deleted).

To avoid this in the future we can query AWS for the ID by security group name instead of hardcoding it. I will do this in another changelist since I want to quickly unblock anyone who needs a AWS preflight.

Change-Id: I9d697ac59bedd7341f7e88c859616818afa0be9b