Make Azure Enumeration Tests more reliable
Recently a number of Azure preflight fail because of similar reasons.
The problem is because Azure Enumeration tests count the resources in Azure
then enumerate and then expect local resources to match azure counts retrieved
at the beginning. This is not working correctly if another preflight is
running simultaneously since Azure resources are created/deleted along
the way and counts don't match.

This changelist is fixing: https://jira-hzn.eng.vmware.com/browse/VSYM-5300

This changelist changes the assertion logic to:
1. provision some resources
2. create some stale resources states
3. enumerate
4. check that the provisioned resources were enumerated and exist locally.
5. check that stale resource states were cleaned up.

List of recently failed preflights:
https://symphony-ci.adlbg.eng.vmware.com/job/photon-model-azure-deploy/2384/console
https://symphony-ci.adlbg.eng.vmware.com/job/photon-model-azure-deploy/2387/console
https://symphony-ci.adlbg.eng.vmware.com/job/photon-model-azure-deploy/2386/console
https://symphony-ci.adlbg.eng.vmware.com/job/photon-model-azure-deploy/2385/console
https://symphony-ci.adlbg.eng.vmware.com/job/photon-model-azure-deploy/2383/console

Change-Id: I96ff2c672fd34df21758eaf05bf464978c76a723