Disk type required annotation change
Removed the disk type required annotation for the disk state, as it is not
mandatory to fill in for provisioning a disk. Also even if they specify as
say SSD, it is upto the provider to satisfy the requirement. Since it is
more of a requirement which can be captured in the constraint field in
DiskState, making as a top level mandatory field is not needed. Hence
removed it after discussion.

Change-Id: I560a6c7084f4901e64c8a24c409a018d1c553f2d