Ensure all the fields populated for ComputeDescription are as well poopulated for ComputeState
This is a first step towards removing the creation of StorageDescriptions during enumeration.
Creating Description for each VM and not deleting it afterwards, is causing problems for
Symphony implementation. That is why, the desicion is, to create only States.

Due to this change, only few additional fields in the State are populated.
Some submitted new fields in ComputeState are deleted, because of a missunderstanding
in the previous CL, that they are needed. Thus, the change is totally not related to any of the
logic, and is not breaking anything.

Validated through the preflight tests, covering enumeration scenarios.

Change-Id: I5587e0c64b7115db62339a8cd3636cbbe39b3126