Add IpHelper class
Utility methods for dealing with IP addresses.
The class is originally created in photon-controller project:
https://github.com/vmware/photon-controller/blob/37d3cafe21b2344c9d2a15387f2b19c075b5f0cf/controller/common/src/main/java/com/vmware/photon/controller/common/IpHelper.java
but is modified to get rid of guava dependency.

Tests are also included (reworked to use JUnit only)

I intent to use this class in dynamic CIDR allocation,
but wanted to check it in lower in the stack since photon-model
 adapters might benefit from its functionality.

Change-Id: I3ae9cc651cbf7252ec37e016738dee8e1ebd2e54