[VCOM-879] [Azure] Set ComputeState.id before starting to provision VM.
During Azure VM provisioning if enumeration is triggered then the
VM that is in created state results in a new ComputeState being created.

When the actual provisioning succeeds we end up with two
ComputeStates describing the same VM.

With this change the VM id is generated and updated *before*
the actual VM provisioning. If enumeration is triggered while
the machine is in provisioning state then the already existing
compute state is updated rather than a new state being creted.

This should fix JIRA VCOM-879 -
https://jira-hzn.eng.vmware.com/browse/VCOM-879

Change-Id: I10d50ee8039b5a13cd2edb5b96b30acbc0a29e07