Filter ResourceGroups by type
A ResourceGroup for Azure contains a property in the ComputeProperties,
specifying its type.
key: ComputeProperties.RESOURCE_TYPE_KEY
value, valid for azure resource group: ResourceGroupStateType.AzureResourceGroup

Enhancing the algorithm for assigning  a ResourceGroup to a SecurityGroup
and Network to filter the provided list of ResourceGroups by type, and to
choose the 1st one from the ones of type
ResourceGroupStateType.AzureResourceGroup.

Change-Id: I0ab3a744195b848efea2467755fcfb094d214bc6