[D2O] support for ResourceOperationSpec extensions
Introduce new field in ResourceOperationSpec to enable contributors to enhance ResourceOperationSpec with additional data, like UI related aspects.
The extensions content is contract between the contributor (e.g. custom UI for resource operations) and the consumers of the meta-data related to the resource operations (e.g. the application which is going to present/visualize the resource operation)
implement toString, hashCode and equals for SchemaField

Testing Done: junit

Change-Id: Id7da3724023502243d7fea214e5795ec0ada4f4b