Add AWSInstanceTypeService
The service returns a list of instance types supported by AWS.
For now we return the values from the AWS SDK enum. If needed
we can easily enhance the representation by adding more elaborate
description to each of the values (descriptions can be stored
in a separate file).

The data model that is returned will be shared with the rest
of the adapters (Azure & VIO). If needed the model can also
be extended with other properties (num CPUs, Memory, etc).

Change-Id: I35d70a940713bf67e61859c4911ebb368850d9f1