Get device backing information for machine network card
Issue: When we are creating network backing info for machine NIC, there are lots of permutations. I wanted to add a permutation for NSX-V but realized that code need some refactoring.
Here I created a factory to return device backing info for the network card. The way I am deciding is by the custom property on a resource. It appears that every time a network is provisioned or collected, we have this custom property set on subnet or network. I am relying on these custom properties to determine which backing info needs to be created for the respective network card. Please review this approach.
Another note: I am using queryCconfigTarget in the scenario when NSX-V logical switch needs to be connected. NSX-V doesn't provide the switchUUID for the port group and that needs to be queried at run time

Bug Number:
Testing done: added unit test for the factory. IT test will cover the provisioning. I'll run the manual tests.

Change-Id: Ic7dc99ef2205c13ed6f136be3a3890f93901d2f9