Schema changes for Snapshots as Day 2 operations:
- Snapshots are hierarchical in nature with tree structure, extended snapshotstate
to contain link to the parent snapshot (if any). To identify parent snapshotstate
to form the hierarchy during create snapshot operation, added a new flag
to identify a snapshot state as the current state for a compute. At any point amongst all
snapshots at most only one snapshot may have this flag set to true.

- Compute day 2 operations - delete and revert snapshot are enabled only if compute has
at least one snapshot available. Snapshot contains a link to Compute but not vice versa.
New custom property defined in ComputeProperties as a flag (yes or no) to indicate
whether a snapshot exists for given compute or not.

Change-Id: Ied84d7bb0e93211acaf45ee48c588589f6463668