[VCOM-1296] Fix NullPointerException in AzureComputeHostStorageStatsGatherer
NullPointerException in AzureComputeHostStorageStatsGatherer when cleaning up
azure rest client when there is no storage account in the subscription.
The NPE is in the following line:
AzureUtils.cleanUpHttpClient(this.restClient.httpClient());
Root cause: There is no storage account and this.restClient is not initialized in this case.
The the finish method will fail. This is preventing azure enumeration (networks, NSG, etc).

https://jira-hzn.eng.vmware.com/browse/VCOM-1296

Change-Id: Id3b0baabd9cb6284a3a70c86abd47e83ada6f249