VSYM-7607: Set type tag to Azure computes
As part of a larger effort around resource tags
and filtering, we are adding internal type tags
to all types of resources. In this commit we are
adding the internal tag to azure computes.

Also I am fixing a related issue found in the
enumeration where tenantLinks were not set
properly before creating the external tags.

Testing: Unit tests and manual testing done.

Change-Id: I07704fbb8058101616580f685f5b63f15c533fc2