Fixing the name of azure blob resources with trailing "/" in URI
There was a bug in resource creation for some azure blobs that have
a trailing "/" in the URI. It caused the resource name to be empty string.
This was happening due to the function we used from xenon utils.

A simple fix is included to remove the trailing "/" from the URI
of azure blobs so that a resource with a correct name can be constructed.

* Added a utility function to parse the name of azure blob.
* Added a unit test to verify the fix.

Testing: manually tested and verified with the added unit test.

Change-Id: Ie5bb09e50b51118062d24b6eac8a6a58c43dc85b