Fix the azure-client-* version to latest release 1.0.4
azure-client-authentication and azure-client-runtime versions are
defined as [1.0.1,1.1.0) in azure-parent. This leads to build that cannot
be reproduced:
  1. When version 1.0.5 is released it will match automatically and next
re-builds will start using it while the previous re-builds of same changeset
will use 1.0.4.
  2. In provisioning-service build snapshots repo is enabled, as result
mvn resolves the dependency to 1.0.5-SNAPSHOT (the newest version in the repos).
This breaks the product as the snapshot version has some issues.

JIRA: VCOM-1545

Change-Id: Ic6ef9046c6c0c0edae1f6085db16f48dfba6b563