VRA-11503 IP Address allocation through isolated network
1. Added field names in SubnetRangeState to be used by queries
2. Modified validation of IPs in range to include the first and last IP in the range as valid.
This was the behaviour in vRA.
Range validation is done by SubnetUtils class from apache commons-net package.
To include the edge IPs in the range, setInclusiveHostCount has to be set to true, which is not supported in apache-commons-net 2.0.
Therefore this change includes bumping commons-net package version to 3.5, same version as used in photon-model/photon-model/pom.xml.
Tested as part of a bigger change to the provisioning service.

Change-Id: Iacce3c245b476ac772f02dffed2f7c6166417205