WIP:VSYM-8581 :Adding logic to map cloud account to a single compute host.
- Earlier there was a 1:1 mapping between an endpoint and a
  compute host.
- Modifying the flow in the EndpointAllocatinTaskService to
  check if a compute host already exists for a given cloud
account.
- If an existing compute host/compute description is found
  then that is used while the endpoint if being configured
as opposed to creating a new one each time.
- This work has been done for AWS. And a placeholder method
  has been added for Azure. The vSphere side of the adapters
may have to be updated to avoid errors due to the contract
change.

Next steps :

- Add a unit test for endpoint allocation task service that
  configures 2 endpoints pointing to the same aws account.
(Did end to end testing for this flow in Symphony. Needs to
be done in a unit test in photon-model as well)

- Solidify the understanding of the Azure workflow to check
  for account uniqueness and potentially refactor the
uniquess check as a common method in the
EndpointAdapterUtils as part of the handleEndpointRequest().
For now have added the functionality to the individual
adpaters.

- Extend the testing work to check how this would work for
  Azure and vSphere endpoints and ensure we update the
adapters to adhere to the new contract.

Change-Id: Ic220d66fb0c90204675c19af2040fe33e08e735e