VSYM-10505 Remove protocol name from Azure resource IDs
Description:
For Azure resource groups/disks, ID field starts with http/https.
We need to start removing protocol as it's not really a part of ID
and adding the protocol can result in duplicate resources due to
inconsistencies between http/https while enumerating.

We also need a transformation service to remove the protol from
existing resources in the environment. Jira here
https://jira-hzn.eng.vmware.com/browse/VSYM-10542

Change-Id: Ic350b01a79d40076a6049cd82fba2b24898cf116