Use the same authz context for initial and periodic scheduled task execution
Initial execution used to use the user context while periodic ones
were using the system context. Fixed that to use the same context in
both cases.

Also added an option to avoid the delay on initial execution. This is
often due to an user action and executing the task immediately is more
predictable. The option defaults to FALSE for backward compatibility.

Change-Id: Ie054dc68ee35f339945858a43e9c72797f1a5933