[AWS] EBS enumeration is creating duplicate diskstates for the already       existing ones.
-This is because enumeration is not finding the diskstate corresponding
to a disk on aws. The reason is that Enumeration uses storageType field
as the filter for locating existing diskstates and it is expecting storageType
to be a first class field of DiskState Where as diskstates created by
provisioning-service saves storageType as a customProperty. As a result,
the search was not returning the diskstates created by provisioning-service

Change-Id: Idbd94a86df4a004d7cf0c2509a052d593f333b62