CV-340 Create IPAddressState for the explicit IP address after validation
During allocation when IP and subnet are specified, the IP will become managed via IpAddressState object if there is a subnet range that match the given IP and the IP is available.
If there are no subnet ranges at all - the task completes silently. In this case the IP will be used but not managed by IPAM.
If there are subnet ranges but none match the IP or the IP is already allocated to a different resource the task fails.

Added unittest and tested e-2-e from cadenza for machine.
More testing required for load balancer VIP.

Change-Id: Iff994a28f5a6ec16095999475018e0ec04c7a37b